Explain that containers and segues don't always just work
The problem is that segues such as FLComposeViewControllerSegue or
the UINavigationController push segue expect the parent view
controllers to be the containers, and this can't work if you're
using nested containers (one container would be the parent, while
the other would be the grandparent).

We could make FLComposeViewControllerSegue look for any ancestor
compose view controller, but then unwind segues wouldn't work
correctly because segueForUnwinding... isn't forwarded to the
parent in a container view controller.

If you want to mix and match container view controllers, I would
recommend connecting an IBAction and looking for a compose view
controller ancestor, then presenting and dismissing as needed.
The navigation view controller would be a child of the compose
view controller, so you could use push segues in the storyboard.