Rename bad method names
Xcode now warns that the present/dismiss methods we used are
deprecated. Actually, Xcode is right in the sense that you're
expected to override methods when you want to replace them, so if
they're deprecated in the base class, they're also deprecated in
its descendants. This is not what was intended in the first place,
overriding those methods was a bad, bad idea.