Layout the root view container manually
We can't rely on autoresizing hints for the root view container
because the transform affects its frame. So we reset it and apply
the transform again after the view layouts itself. We can't try to
reset the frame and transform during the animation, though, so we
have to keep track of that.