Set the modal view frame before the appearance transition
That way the new frame is available for the appearance methods on
the modal view. Note how the modal view container frame was also
incorrectly set in loadView, since it's reset before we present
the modal view. Also note that we were resetting the modal view
frame with the wrong value, since the modal view container frame
hadn't yet been set.