Reset the root container effect after the presentation
We need to reset the container effect after the presentation,
otherwise the frame will be wrong by the time we dismiss. If the
rotation happened during the presentation animation, we don't reset
it in viewDidLayoutSubviews because that would halt the animation.
Instead, we force the reset after the presentation animation.

Note that viewDidLayoutSubviews is called after the dismiss
animation, but not after the presentation animation, which is why
we don't need to call reset after the dismiss animation. This is
probably because the dismiss completion block changes the view
hierarchy, which in turn invalidates the layout.